requirements.txt
# === Core environment ===
python-dotenv==1.0.1
redis==5.0.8
numpy==1.26.4
sounddevice==0.4.6
webrtcvad==2.0.10
faster-whisper==1.0.1
rapidfuzz==3.9.6
openai==1.46.0

# === Async backend / API layer ===
fastapi==0.115.2
uvicorn==0.31.1
pydantic==2.9.2
python-multipart==0.0.9

# === Speech & TTS ===
piper-tts @ git+https://github.com/rhasspy/piper.git@master
# If you’re using the CLI binary instead of Python bindings, remove this line.

# === Misc utilities ===
requests==2.32.3
aiofiles==24.1.0
jinja2==3.1.4
httpx==0.27.2

# === Optional debugging / dev tools ===
rich==13.9.2
watchdog==4.0.1
colorama==0.4.6
